DjangoRestAPI/requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
mssql-django==1.1
pyodbc==4.0.32
pytz==2021.3
sqlparse==0.4.2
